Implement move counter
Every time a legal move is made (little spider: tableau to tableau,
tableau to homecell, homecell to tableau. golf: tableau to homecell,
stockpile to homecell.) 1 move is added the moves instance variable in
the game class. A .csv file was created to store the high score data
from game play to game play.